added formTitle to own file
